// src/services/types/ReviewFilters.ts
export type ReviewFilters = {
  /**
   * @description Review rating between 1 to 5
   * @minLength 1
   * @maxLength 5
   * @type integer | undefined
   */
  star_rating?: number;
  /**
   * @description If review has response
   * @type boolean | undefined
   */
  has_comment?: boolean;
  /**
   * @description If review has a business reply (reply field = 1)
   * @type boolean | undefined
   */
  has_reply?: boolean;
  /**
   * @description If review has been deleted from Google
   * @type boolean | undefined
   */
  is_deleted?: boolean;
  /**
   * @description Page number
   * @type number | undefined
   */
  page?: number;
  /**
   * @description Results per page. Default is 10. Maximum is 50
   * @maxLength 50
   * @default 10
   * @type number | undefined
   */
  per_page?: number;
};