// src/services/types/ReviewReplyRequest.ts
export type ReviewReplyRequest = {
  /**
   * @description Reply text content
   * @minLength 1
   * @maxLength 1000
   * @type string
   */
  text: string;
  /**
   * @description Whether the reply should be visible to the public
   * @default true
   * @type boolean | undefined
   */
  isPublic?: boolean;
};